Funcionalidad para iniciar sesión.

+++ js/categories.js
@@ -5,7 +5,6 @@ let currentCategoriesArray = [];
 let currentSortCriteria = undefined;
 let minCount = undefined;
 let maxCount = undefined;
-let userEmail = JSON.parse(localStorage.getItem("userPersonalInfo")).email;
 
 function sortCategories(criteria, array){
     let result = [];
@@ -97,11 +96,20 @@ document.addEventListener("DOMContentLoaded", function(e){
         }
     });
     
-    document.getElementById("profile").innerHTML = `${userEmail}`;
-
-    document.getElementById("logout").addEventListener("click", function(){
-        localStorage.clear();
-    });
+    if(JSON.parse(localStorage.getItem("userPersonalInfo")) !== null){
+        let userEmail = JSON.parse(localStorage.getItem("userPersonalInfo")).email;
+        document.getElementById("nav-profile").innerHTML = `
+        <a class="nav-link dropdown-toggle" id="profile" role="button" data-bs-toggle="dropdown" aria-expanded="false">${userEmail}</a>
+        <ul class="dropdown-menu" aria-labelledby="profile">
+            <li><a class="dropdown-item" href="cart.html"><i class="fa fa-shopping-cart"></i> Mi carrito</a></li>
+            <li><a class="dropdown-item" href="my-profile.html"><i class="fa fa-user"></i>  Mi perfil</a></li>
+            <li><a class="dropdown-item" href="login.html" id="logout"><i class="fa fa-sign-out"></i> Cerrar sesión</a></li>
+        </ul
+        `
+        document.getElementById("logout").addEventListener("click", function(){
+            localStorage.clear();
+        });
+    };
 
     document.getElementById("sortAsc").addEventListener("click", function(){
         sortAndShowCategories(ORDER_ASC_BY_NAME);
