+++ js/product-info.js
@@ -1,4 +1,4 @@
-let userEmail = JSON.parse(localStorage.getItem("userPersonalInfo")).email;
+let userEmail;
 let prodID = localStorage.getItem("prodID");
 let cartArticles = JSON.parse(localStorage.getItem("cartArt"));
 let productInfo = undefined;
@@ -154,16 +154,18 @@ function showProduct(prod){
     document.getElementById("product-info").innerHTML = productInfo;
     document.getElementById("related-products").innerHTML = relatedProducts;
     document.getElementById("btn-add-to-cart").addEventListener("click", function(){
-        let amount = parseInt(document.getElementById("product-amount").value);
-        addToCart(prod,amount);
-        document.getElementById("add-comment").innerHTML += `
-        <div class="alert alert-success alert-dismissible fade show" role="alert">
-            ¡El producto se agregó correctamente!
-            <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-        </div>
-        `
+        if(userEmail !== undefined){
+            let amount = Math.round(document.getElementById("product-amount").value);
+            addToCart(prod,amount);
+            document.getElementById("cart-alert").classList.add("alert-success");
+            document.getElementById("cart-alert").classList.add("show");
+            document.getElementById("msg-cart-alert").innerHTML = `¡El producto se agregó correctamente!`
+        } else {
+            document.getElementById("danger-alert").classList.add("alert-danger");
+            document.getElementById("danger-alert").classList.add("show");
+            document.getElementById("msg-danger-alert").innerHTML = `Debes iniciar sesión para agregar productos al carrito.`
+        };
     });
-
     for(let i = 0; i < imgArray.length; i++){
         document.getElementById(`img${i}`).addEventListener("mouseover",function(){
             document.getElementById("main-img").src = `${imgArray[i]}`;
@@ -188,11 +190,20 @@ document.addEventListener("DOMContentLoaded",function(){
         };
     });
 
-    document.getElementById("profile").innerHTML = `${userEmail}`;
-
-    document.getElementById("logout").addEventListener("click", function(){
-        localStorage.clear();
-    });
+    if(JSON.parse(localStorage.getItem("userPersonalInfo")) !== null){
+        userEmail = JSON.parse(localStorage.getItem("userPersonalInfo")).email;
+        document.getElementById("nav-profile").innerHTML = `
+        <a class="nav-link dropdown-toggle" id="profile" role="button" data-bs-toggle="dropdown" aria-expanded="false">${userEmail}</a>
+        <ul class="dropdown-menu" aria-labelledby="profile">
+            <li><a class="dropdown-item" href="cart.html"><i class="fa fa-shopping-cart"></i> Mi carrito</a></li>
+            <li><a class="dropdown-item" href="my-profile.html"><i class="fa fa-user"></i>  Mi perfil</a></li>
+            <li><a class="dropdown-item" href="login.html" id="logout"><i class="fa fa-sign-out"></i> Cerrar sesión</a></li>
+        </ul
+        `
+        document.getElementById("logout").addEventListener("click", function(){
+            localStorage.clear();
+        });
+    };
 
     document.getElementById("add-comment").addEventListener("submit",function(event){
         event.preventDefault();
@@ -204,7 +215,9 @@ document.addEventListener("DOMContentLoaded",function(){
             document.getElementById("textArea").value = "";
             document.getElementById("select-score").value = 0;
         } else {
-            alert("Por favor ingrese un comentario y un puntaje antes de enviar.")
+            document.getElementById("warning-alert").classList.add("alert-warning");
+            document.getElementById("warning-alert").classList.add("show");
+            document.getElementById("msg-warning-alert").innerHTML = `Por favor ingrese un comentario y un puntaje antes de enviar.`
         };
     });
